Add a generic delivery status healthcheck
